refactor: move sending messages to channel in its own handler and add actions listener registration
