+++ src/slack/listeners/actions.ts
@@ -1,0 +1,6 @@
+import { App } from '@slack/bolt';
+import { handleNotifyAllAction } from '../handlers/actions/notifyAllChannels';
+
+export const registerActionListeners = (slackApp: App) => {
+  handleNotifyAllAction(slackApp);
+};
